third_party/WebKit/LayoutTests/fast/scrolling/unscrollable-subpixel-overflow-mouse-drag.html

<!DOCTYPE html>
<style>
body {
  zoom: 175%;
  margin: 0;
}
#container {
  overflow-y: scroll;
  width: 200px;
  height: 200px;
}
#content {
  background-color: green;
  width: 100px;
  height: 200.3px;
}
</style>
Tests scroll behavior when an unscrollable scrollbar thumb is dragged by mouse.
Passes if neither the thumb or the content changes.
<div id="container">
  <div id="content"></div>
</div>
<script src="../../resources/js-test.js"></script>
<script>
if (window.eventSender) {
  onload = function() {
    var container = document.getElementById('container');
    eventSender.mouseMoveTo(340, 200);
    eventSender.mouseDown();
    eventSender.mouseMoveTo(340, 250);
    eventSender.mouseUp();
    shouldBe("container.scrollTop", "0");
  };
}
</script>
